+++ tests/infrastructure/projections_test.py
@@ -4,10 +4,10 @@ from uuid import UUID
 
 import pytest
 
-from sharedkernel.domain.errors import UnhandledEventType
 from sharedkernel.domain.events import DomainEvent
+from sharedkernel.domain.exceptions import UnhandledEventType
 from sharedkernel.infrastructure.data import DataModel
-from sharedkernel.infrastructure.errors import EventOutOfSequence
+from sharedkernel.infrastructure.exceptions import EventOutOfSequence
 from sharedkernel.infrastructure.projections import Projection, Projector
